Changed player, created gem, changed health system
I changed the player knockback to be an arc rather than a teleport. I created a gem pickup to give the player a direction. The gem pickup awards an amount of points set by the designer. I removed "heart pieces" from the health system and it now only uses lives instead of partial heart pieces that make a whole life.